requirements-prod.txt
# Production Requirements
# Includes all base requirements plus production-specific packages

# Base requirements
-r requirements.txt

# Production WSGI Server
gunicorn==21.2.0
gevent==23.9.1

# Environment variable management
python-dotenv==1.0.0

# Rate limiting
Flask-Limiter==3.5.0
redis==5.0.1

# Monitoring and metrics
prometheus-client==0.19.0
sentry-sdk[flask]==1.39.2

# Database (if migrating to PostgreSQL)
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Security
Flask-Talisman==1.1.0  # HTTPS enforcement and security headers
cryptography==41.0.7

# Caching
Flask-Caching==2.1.0

# Request validation
flask-validator==1.4.4

# Health checks
healthcheck==1.3.3

# Structured logging
python-json-logger==2.0.7

# Testing (for CI/CD)
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0

# Code quality
black==23.12.1
isort==5.13.2
flake8==6.1.0
pylint==3.0.3
bandit==1.7.6
safety==2.3.5
